Blueprint setting benchmark
This file is a benchmark file for blueprint settings. The AWS Gen5 model is used to optimize the Gen4 model recommended by the existing CloudEndure platform by default. After the 6th generation model is launched on the AWS platform, you can also modify this file to set the benchmark model of the new generation for blueprint